+++ scalable-payment-processing/src/main/java/io/confluent/kpay/payments/PaymentsConfirmed.java
@@ -1,8 +1,9 @@
 package io.confluent.kpay.payments;
 
-import io.confluent.kpay.ktablequery.*;
 import io.confluent.kpay.payments.model.ConfirmedStats;
 import io.confluent.kpay.payments.model.Payment;
+import io.confluent.kpay.rest_iq.WindowKTableResourceEndpoint;
+import io.confluent.kpay.rest_iq.WindowKVStoreProvider;
 import org.apache.kafka.common.serialization.Serdes.StringSerde;
 import org.apache.kafka.common.utils.Bytes;
 import org.apache.kafka.streams.KafkaStreams;
