+++ scalable-payment-processing/src/main/java/io/confluent/kpay/rest_iq/MicroRestService.java
@@ -1,4 +1,4 @@
-package io.confluent.kpay.ktablequery;
+package io.confluent.kpay.rest_iq;
 
 
 import org.eclipse.jetty.server.Server;
