+++ scalable-payment-processing/src/main/java/io/confluent/kpay/rest_iq/WindowStoreProvider.java
@@ -1,4 +1,4 @@
-package io.confluent.kpay.ktablequery;
+package io.confluent.kpay.rest_iq;
 
 import org.apache.kafka.streams.state.ReadOnlyWindowStore;
