+++ scalable-payment-processing/src/test/java/io/confluent/kpay/rest_iq/MicroRestServiceTest.java
@@ -1,4 +1,4 @@
-package io.confluent.kpay.ktablequery;
+package io.confluent.kpay.rest_iq;
 
 import io.confluent.kpay.payments.model.Payment;
 import io.confluent.kpay.util.GenericClassUtil;
